requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0

# Environment
gymnasium>=0.28.0

# Visualization
plotly>=5.14.0
kaleido>=0.2.1  # For static image export

# Configuration
pyyaml>=6.0

# Logging & Monitoring
tensorboard>=2.12.0
tqdm>=4.65.0

# Data & Analysis
pandas>=2.0.0
scipy>=1.10.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Development
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
